ProjectCard/index.js renamed to ProjectCard.js
