+++ src/components/pages/Portfolio.js
@@ -1,5 +1,5 @@
 import React from "react";
-import ProjectCard from "./ProjectCard/";
+import ProjectCard from "./ProjectCard.js";
 import projects from "./../../projects.json";
 
 const styles = {
